Using package testcontainers
